requirements.txt
torch
torchvision
Pillow
streamlit